Skip rebuild from scratch after cmake is ran (#1636)
By commiting the patch changes, this commit prevents the libraries to be
built from scratch every time cmake is ran. This will save additional
build time.

In addition to this, this commit separates out the different
functionality to make the build system more readable. We can continue to
iterate on this in the future.

Signed-off-by: John Mazanec <jmazane@amazon.com>