Encapsulate dimension, vector data type validation/processing inside Library (#1957)
Introduces a new configuration object KNNMethodConfigContext. 
The KNNMethodContext contains the user provided
information for want their index to be built like. However, it is
missing a few pieces that are defined outside of it. These
pieces are needed to validate the config and actually build the config.

This change Integrates validation of KNNMethodContexts with 
KNNMethodConfigContext and better encapsulates KNNLibrary 
config info in the engine package

Signed-off-by: John Mazanec <jmazane@amazon.com>